requirements.txt
PyQt5
twisted
python-binance
pandas